requirements.txt
flask
flask-limiter
werkzeug
python-dotenv
torch
openai-whisper
rapidfuzz
gunicorn